staffbackend/requirements.txt
Django==5.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
python-decouple==3.8
